# src/FacilityManagement.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FacilityManagement",
    "description": "",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/MaintenanceRequest/{maintenanceRequestId}": {
      "get": {
        "tags": [
          "MaintenanceRequest"
        ],
        "summary": "Get MaintenanceRequest by id.",
        "description": "Get MaintenanceRequest by id.",
        "operationId": "getMaintenanceRequestById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maintenanceRequestId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetMaintenanceRequestByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MaintenanceRequest"
        ],
        "summary": "Delete MaintenanceRequest.",
        "description": "Delete MaintenanceRequest.",
        "operationId": "removeMaintenanceRequest",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maintenanceRequestId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveMaintenanceRequestResponse"
            }
          }
        }
      }
    },
    "/api/MaintenanceRequest": {
      "get": {
        "tags": [
          "MaintenanceRequest"
        ],
        "summary": "Get MaintenanceRequests.",
        "description": "Get MaintenanceRequests.",
        "operationId": "getMaintenanceRequests",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetMaintenanceRequestsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "MaintenanceRequest"
        ],
        "summary": "Create MaintenanceRequest.",
        "description": "Create MaintenanceRequest.",
        "operationId": "createMaintenanceRequest",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateMaintenanceRequestRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateMaintenanceRequestResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "MaintenanceRequest"
        ],
        "summary": "Update MaintenanceRequest.",
        "description": "Update MaintenanceRequest.",
        "operationId": "updateMaintenanceRequest",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateMaintenanceRequestRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateMaintenanceRequestResponse"
            }
          }
        }
      }
    },
    "/api/MaintenanceRequest/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "MaintenanceRequest"
        ],
        "summary": "Get MaintenanceRequest Page.",
        "description": "Get MaintenanceRequest Page.",
        "operationId": "getMaintenanceRequestsPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetMaintenanceRequestsPageResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Address": {
      "type": "object",
      "properties": {
        "street": {
          "type": "string",
          "readOnly": true
        },
        "unit": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "city": {
          "type": "string",
          "readOnly": true
        },
        "province": {
          "type": "string",
          "readOnly": true
        },
        "postalCode": {
          "type": "string",
          "readOnly": true
        }
      }
    },
...
    "CreateMaintenanceRequestRequest": {
      "type": "object",
      "properties": {
        "maintenanceRequest": {
          "$ref": "#/definitions/MaintenanceRequestDto"
        }
      }
    },
    "CreateMaintenanceRequestResponse": {
      "type": "object",
      "properties": {
        "maintenanceRequest": {
          "$ref": "#/definitions/MaintenanceRequestDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "GetMaintenanceRequestByIdResponse": {
      "type": "object",
      "properties": {
        "maintenanceRequest": {
          "$ref": "#/definitions/MaintenanceRequestDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetMaintenanceRequestsPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MaintenanceRequestDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetMaintenanceRequestsResponse": {
      "type": "object",
      "properties": {
        "maintenanceRequests": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MaintenanceRequestDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "MaintenanceRequestCommentDto": {
      "type": "object",
      "properties": {
        "maintenanceRequestCommentId": {
          "format": "uuid",
          "type": "string"
        },
        "maintenanceRequestId": {
          "format": "uuid",
          "type": "string"
        },
        "body": {
          "type": "string"
        },
        "created": {
          "format": "date-time",
          "type": "string"
        },
        "createdById": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "MaintenanceRequestDigitalAssetDto": {
      "type": "object",
      "properties": {
        "maintenanceRequestId": {
          "format": "uuid",
          "type": "string"
        },
        "digitalAssetId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "MaintenanceRequestDto": {
      "type": "object",
      "properties": {
        "maintenanceRequestId": {
          "format": "uuid",
          "type": "string"
        },
        "requestedByProfileId": {
          "format": "uuid",
          "type": "string"
        },
        "requestedByName": {
          "type": "string"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "address": {
          "$ref": "#/definitions/Address"
        },
        "phone": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "unattendedUnitEntryAllowed": {
          "type": "boolean"
        },
        "receivedByProfileId": {
          "format": "uuid",
          "type": "string"
        },
        "receivedByName": {
          "type": "string"
        },
        "receivedDate": {
          "format": "date-time",
          "type": "string"
        },
        "workDetails": {
          "type": "string"
        },
        "workStarted": {
          "format": "date-time",
          "type": "string"
        },
        "workCompleted": {
          "format": "date-time",
          "type": "string"
        },
        "workCompletedByName": {
          "type": "string"
        },
        "unitEntered": {
          "$ref": "#/definitions/UnitEntered"
        },
        "requestedByProfile": {
          "$ref": "#/definitions/ProfileDto"
        },
        "status": {
          "$ref": "#/definitions/MaintenanceRequestStatus"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MaintenanceRequestCommentDto"
          }
        },
        "digitalAssets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MaintenanceRequestDigitalAssetDto"
          }
        }
      }
    },
    "MaintenanceRequestStatus": {
      "format": "int32",
      "type": "integer"
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "ProfileDto": {
      "type": "object",
      "properties": {
        "profileId": {
          "format": "uuid",
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      }
    },
...
    "RemoveMaintenanceRequestResponse": {
      "type": "object",
      "properties": {
        "maintenanceRequest": {
          "$ref": "#/definitions/MaintenanceRequestDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "UnitEntered": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    },
...
    "UpdateMaintenanceRequestRequest": {
      "type": "object",
      "properties": {
        "maintenanceRequest": {
          "$ref": "#/definitions/MaintenanceRequestDto"
        }
      }
    },
    "UpdateMaintenanceRequestResponse": {
      "type": "object",
      "properties": {
        "maintenanceRequest": {
          "$ref": "#/definitions/MaintenanceRequestDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}