# src/FacilityManagement.App/src/app/@api/models/create-maintenance-request-response.ts
/* tslint:disable */
import { MaintenanceRequestDto } from './maintenance-request-dto';
export interface CreateMaintenanceRequestResponse {
  maintenanceRequest?: MaintenanceRequestDto;
  validationErrors?: Array<string>;
}