# src/FacilityManagement.App/src/app/@api/models/get-maintenance-requests-page-response.ts
/* tslint:disable */
import { MaintenanceRequestDto } from './maintenance-request-dto';
export interface GetMaintenanceRequestsPageResponse {
  entities?: Array<MaintenanceRequestDto>;
  length?: number;
  validationErrors?: Array<string>;
}