# src/FacilityManagement.App/src/app/@api/models/maintenance-request-comment-dto.ts
/* tslint:disable */
export interface MaintenanceRequestCommentDto {
  body?: string;
  created?: string;
  createdById?: string;
  maintenanceRequestCommentId?: string;
  maintenanceRequestId?: string;
}