# src/FacilityManagement.App/src/app/@api/models/remove-maintenance-request-response.ts
/* tslint:disable */
import { MaintenanceRequestDto } from './maintenance-request-dto';
export interface RemoveMaintenanceRequestResponse {
  maintenanceRequest?: MaintenanceRequestDto;
  validationErrors?: Array<string>;
}