# src/FacilityManagement.App/src/app/@api/models/update-maintenance-request-response.ts
/* tslint:disable */
import { MaintenanceRequestDto } from './maintenance-request-dto';
export interface UpdateMaintenanceRequestResponse {
  maintenanceRequest?: MaintenanceRequestDto;
  validationErrors?: Array<string>;
}